Attempted fix on edge bugs for highlighting
Highlighting for the line of sight works better, but still isn't
perfect. Combined the algorithm class with the main drawer so loading up
a gallery for part 1 will also place guards at positions based on what
the algorithm calculates. Bare in mind the algorithm will most likely
fail on the more complex galleries due to existing bugs. Also commented
out the drawing of lines for the guards to speed the program up a
little.